refactor: try to separate curl call onto each corresponding files, update cmakelists
