+++ CMakeLists.txt
@@ -32,10 +32,21 @@ else()
   FetchContent_MakeAvailable(json)
 endif()
 
+include_directories(
+    ${PROJECT_SOURCE_DIR}/include
+    ${PROJECT_SOURCE_DIR}/src
+)
+
+file(GLOB all_SRCS
+    "${PROJECT_SOURCE_DIR}/include/*.h"
+    "${PROJECT_SOURCE_DIR}/include/*.hpp"
+    "${PROJECT_SOURCE_DIR}/src/*.cpp"
+    "${PROJECT_SOURCE_DIR}/src/*.c"
+)
 
 # Create an executable
 add_executable(${PROJECT_NAME}
-    src/main.cpp
+    ${all_SRCS}
 )
 
 # Linking libraries
