+++ include/coingecko.h
@@ -1,0 +1,13 @@
+#include <iostream>
+
+#include <curl/curl.h>
+
+#include <cstdlib>
+
+#include "nlohmann/json.hpp"
+
+#include <dpp/dpp.h>
+
+namespace gecko {
+void fetch(dpp::slashcommand_t event);
+}
